+++ src/bufio.h
@@ -13,7 +13,7 @@ ssize_t bufio_readline(struct bufio *self, size_t *line_offset);
 ssize_t bufio_read(struct bufio *self, size_t count, size_t *buf_offset);
 char * bufio_offset2ptr(struct bufio *self, size_t offset);
 size_t bufio_ptr2offset(struct bufio *self, char *ptr);
-ssize_t bufio_sendfile(struct bufio *self, int fd, off_t *off, int filesize);
+ssize_t bufio_sendfile(struct bufio *self, int fd, off_t *off, size_t filesize);
 ssize_t bufio_sendbuffer(struct bufio *self, buffer_t *response);
 
 #endif /* _BUFIO_H */
